Move front page cover to standalone directory
